Make re-add usernames be in the canonical casing
Ran into a bug where the non-canonical casing of a username was
re-added, and then kicked the following week because of case-sensitive
username comparisons.

Case-sensitive username comparisons absolutely should not be happening,
but this is a bandaid fix that will at least make it harder to run into
issues with case-sensitive comparisons.